Further Polishing the game
There was a little error in the previous version of the game. It was
working when I submitted but I think due to one of a archived online
resource, the platform.js, was lost. So this version is mainly a change
for that purpose, and with little other twists for thursday’s in class
demonstration. Please do not use this commit for grading hw6 if this
would be considered as a late submission. I did submit hw on time, and
if there is a problem with previous version, it because one of a
reference file is no longer available.